requirements.txt
Django==4.1.7
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
django-import-export==3.1.0
drf-extra-fields==3.4.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
django-filter==23.1







